ajuste recomendacoes

+++ rentGames/src/main/kotlin/br/com/mekylei/rentgames/model/InfoGameJson.kt
@@ -1,6 +1,8 @@
 package br.com.mekylei.rentgames.model
 
-data class InfoGameJson(val titulo: String, val capa: String, val preco: Double, val descricao: String) {
+import java.math.BigDecimal
+
+data class InfoGameJson(val titulo: String, val capa: String, val preco: BigDecimal, val descricao: String) {
 
     override fun toString(): String {
         return "InfoGameJson(titulo='$titulo', capa='$capa', preco='$preco', descricao='$descricao')"
